move clear memory to utils memory dir
